Changing everything. This is the first working version of the doubly linked list being moved over to the new idea of "kinds." All dllist functions now operate via just "list" functions (i.e., dllist_push will now be simply list_push).
This is a major update, including the removal of node operations on lists (as it
exposes the internal structure of the list). In favor of good practice, the only
way to interact with data in the list now is via the element pointers (void *)
so that the internal list structure can be abstracted away. This allows for the
kind type to be used to create different implementations of lists that are
elegantly interoperable.

Also, lists now support concatenation via the list_concat() and
list_concat_into() functions. Although not completely implemented (conglomerate
list functionality is missing), it is functional. A future update will introduce
conglomerate lists.

The test code has not yet been updated. This will happen in the next revision,
as the build system has been moved to Visual Studio 2010.